+++ R/RcppExports.R
@@ -5,8 +5,8 @@ raster_info_cpp <- function(pszFilename) {
     .Call('_vapour_raster_info_cpp', PACKAGE = 'vapour', pszFilename)
 }
 
-raster_io_cpp <- function(filename, window, band = 1L) {
-    .Call('_vapour_raster_io_cpp', PACKAGE = 'vapour', filename, window, band)
+raster_io_cpp <- function(filename, window, band = 1L, resample = "NearestNeighbour") {
+    .Call('_vapour_raster_io_cpp', PACKAGE = 'vapour', filename, window, band, resample)
 }
 
 #' Layer names